Fix insertion of space in J command
xvi used to insert 2 spaces between lines if the first ended in space.

This commit fixes that bug and conforms it to the original vi and nvi which,
when joining two lines, strip all space characters from the start of the second,
preserve spaces at the end of the first and then,
if the stripped second line doesn't begin with ')', insert a space.

vim instead, joining "2" to " )" leaves a space, whereas the others don't.