Position the cursor after Join operations.
When joning with the J command, the cursor should be left on the first
character after the first of the joined lines (which is the
space that we inserted if we did, or the first char of the second line
if we didn't).

We also make the :j command leave the cursor at the start of the resulting
line. The original "vi" usually does this but sometimes leaves the cursor
where it was, apparently at random.