Remove the last 3 compiler warnings
GCC complained about dup()'s value being ignored, which is harmless as
it is sure to succeed, so we just assign it to a junk variable.

The other one was "fdopen: return value of int cast to pointer" which is
bizarre because xvi.h includes stdio.h, which should declare fdopen.
Whatever the cause, if it thinks fdopen returns an int that could be
damaging on 64-bit systems which have 32-bit int.
Whatever the cause, adding an explicit type declaration sorts it out,