First part of fixing X-window resizing: find TIOCGWINSZ
xvi was receiving the SIGWINCH but not acting on it on Linux
because the TIOCGWINSZ was not defined. This fixes that.
It now resizes immediately the first time you resize the window
but from then on, only when you press a character bcos read()
is no longet interrupted by the signal SIGWINCH.