Marge "Summary of Changes" document into man page and eliminate it
summary.ms and xvi.1 had a lot of text in common, resulting in
having to maintain both of the in sync. This commit merges the
extra bits and inproved text from summary.ms into the manual page
and deletes the summary document.

It also removes the build dependency "groff" from the debian package,
only required to format a troff document into a text file.