If tags file was missing but appears, read it.
A common workflow is to search for a tag, discover that the tags file
does not exist, suspend the editor, make tags, resume the editor and
retry the tag search. Previously xvi only tried once to read the tags
file. Now it retries on each tag search until it finds one.