Use select() to check for pending keyboard input.
The method for setting read timeouts for input on termios
compatible systems was to adjust the VMIN and VTIME members
of the termios c_cc array, then reset the state. The upshot of
relying on this to cause read() timeout is that in practice, it
doesn't seem to work as expected: read() blocks, and if SIGINT is
handled, then the signal seems to get lost.

This changed removes the termios timeout code and simply uses
select() to check for pending input. Remove input_timeout() function
and simply set current_timeout from inch().