For ^C, add keystroke and add map char.
Previously, for ^C handling in CMDLINE state, the ESC key was
poked into the cmd_input() routine and let that cancel the
cmdmode. Better is to simply treat the kbdintr as a normal key
and set CTRL('C') as the current key. This makes CMDLINE state
cancel in a cleaner manner on ^C.