Make updateline update all affected windows.
Before, when updateline did it's thing, it only updated other
windows into the same buffer if the line was longer than the
size of the current cursor line. That logic seems disjoint from
what we should be looking at: is the line to be updated on any
other screen windows and if so, then update them.
Also, for good measure, add a flag for updating lines below that
line so that when changes to the cursor line affect the rest of
the screen, the changes are propigated properly.