Implement ":source" with flexbufs
There used to be a fixed-size 256-char input line buffer permanently
allocated; now there is no limit on the length of lines in the file.
(and it saves 256 bytes of RAM, which was the reason for doing it!)
while costing only 24 extra bytes of compiled code (on Linux 32-bit
GCC 4.9 -O2).