Account for statusline when updating lines.
When I modified the updateline routine to update associated buffer
windows and also update lines after the specified line, I forgot
to account for the statusline. I noticed the statusline corruption
and mistakenly added an update_sline call instead of subtracting
the statusline from the count of lines to be updated.