Add "towers of hanoi solving vi macros" as a test case
The miraculous hanoi macros exercise a variety of motions, yanks and puts
into named buffers and so on. This is also a regression test case for the
preceding commits that enabled hanoi to work.

At present, the hanoi-solving has to finish in under one second.
An alternative would be a timeout clause.