Include timeouts in hanoi test
Previously xvi was given a fixed time (1 sec) to read, initialize
or run the hanoi macros. This changes it to use term_expect's timeout
so the test completes faster if it succeeds and if it goes into an infinite
loop, the test will fail, and on slower machines the test is unlikely to fail
just because it didn't have time to complete (max set to 10 seconds now).