Fix xvAdjustWindows and xvNextDisplayedWindow.
xvAdjustWindows:
Set nw (number of windows) counter to 0 before counting or we
end up with one more count than we have windows.
Redisplay hidden buffer windows when the screen height increases.
Skip over hidden windows during the screen height downsize loop.
When we are setting a window to be hidden, ensure we don't set
curwin to NULL.
Move the position adjustment outside of the downsize conditional
and rewrite to be more succinct.

xvNextDisplayedWindow:
Check for non-zero height window instead of an arbitrary height.
(I realise that we should probably be checking for P_minrows, but
at this time it is possible to have just a statusline).