Revisit and revise ^C handling.
Stop unsetting the kbdintr flag in the main loop; only unset
it when the code actually acts on it.
Make certain that imessage flag gets set when ever the kbdintr
state gets handled.
Ensure that the 'Interrupted' message only gets printed once for
each SIGINT occurence by removing extraneous messages outside of
the xvi_handle_event routine.
Remove code that tries to second guess whether a CTRL('C') is
supposed to be a SIGINT; that should be handled by the system
specific code by converting CTRL('C') to kbdintr before it gets
out in to the general code.