Fix issue 70: y'[Esc]d'm just moves to mark m
If you set a mark with mm, move to a different line and say y'[Esc]d'm
because you were thinking of yanking the lines, then changed your mind to
delete them, classic vi deletes the lines.
xvi just moves the cursor to mark m. This turns out to be because the
code handling ESC didn't clear cmd->cmd_two_char, so the d was gobbled
as the second character of the NOP command that ESC leaves.
Clearing cmd_two_char on ESC fixes this.