Add virterm
There is an alternative implementation of the "tkterm" interface in
"expect"'s "example/virterm", using arrays to keep the screen content
instead of a Tk text widget, thereby removing the need to have X running.

This change imports the guts of it, removing its different (older?)
version of term_expect and the trailing script to interrogate the
Cornell Library catalog.

It doesn't work yet as a replacement for tkterm: work in progress.