Reorganise terminal emulation software to allow use of virterm
Now, test files just "source term", which chooses between tkterm and
virterm, according to environment variable TKTERM, which now contain
their own versions of "term_expect", which are different for each.

"term" also contains the utility routines that used to be in "term_utils",
so there's no need to include that either.