Simplify tests to work round defect in tkterm test harness
If you go
    :map x y
    :map y x
    xaa^C
the x should send it into an infinite loop, the "aa" be ignored
and then thrown away at the ^C. This happens in real life but with
the tkterm harness, the aa gets processed after the ^C, leaving you
in append mode with an "a" on the screen.

This change simplifies the tests interrupting infinite macro recursion
removing the check that they really are in infinite recursion, to avoid
this behaviour. It doesn't seem grave enough or common enough to fix.