Remove "autogrow" parameter
This undocumented parameter (default: on) controls the ability of
windows to make space to split or create new windowa after them by
making other windows smaller. The only thing it does when off is to make
xvi say "Not enough room!" when your window is small instead of doing
what you asked for.

This change makes it always do that.