Implement the "scroll" parameter
Classic vi sets scroll to half the window size when you resize
xvi's screen, and remembers your own setting until the screen is
resized again. For xvi, it makes more sense to scroll ^U and ^D
by half each window's height, so we leave scroll=0 and scroll by
half a window height if it is unset.