Reduce undo history to one level.
There is an undocumented 100-level undo history in xvi,
accessible only via the :undo and :redo commands, not the 'u' key.
It promises to be a memory hog and is of dubious value.

This change removes that and the associated ":redo" command
saving 752 bytes of code and all the RAM it saved changes in.