Fix issue 93 whereby undoing "dl" moves the cursor one place left
If, in "abcde" you move to the "c", delete it with "dl" and undo with "u"
the cursor was restored to ibe sitting on the "b".
This was because the code to leave the cursor on the first char after the
deleted text moved the cursor left, deleted the text and moved it right
so the stored position for undo was one place to the left.
This fixes that.