Added description of proposed workflow using issues/* and BUGS
This outlines a system to track issues with text files in the
source tree instead of a web service, using files in "issues/"
on the master branch and a file "BUGS" in issue-specific branches.

Apart from keeping the issues with the source they refer to,
the system described has the advantage of simplicity.