Improve termcap's automargins and "xn" logic.
It had
    auto_margins = tgetflag("am") && !tgetflag("xn")
which is not right. On trminals with both "am" and "xn", that would leave
auto_margins false, which means that the cursor stays in the last column.

This makes the code conform better to the meaning of those strings.
In particular, it lets us use the fact that, on terminals without "am",
we know that the cursor hangs at the right edge when printing a normal
char in the last column.