Fix issue 15: Add U command
This adds the missing U command to undo all changes made to a line since
it became the current line. Doing "U" once leaves stuff in the normal
undo history to undo the U command but, unlike classic vi and nvi, if
you do it twice you lose the ability to get the messed-up line back.