Improve colour and cursor motion output behaviour
It used to actually move the cursor too much, resulting is sequences of
empty motions and even outputting move-to-column-81 sequences. It was also
turning the output colour on and off to no effect, as well as outputting
cursor motions while in standout mode, which makes garbage on some terminals.

This changes caches desired colour until an actual character if output
and also adds logic to avoid moving the cursor while in standout mode
unless the terminalis known to be OK with that.