Allow interruption of a big file read
If you edited an immense file, xvi just hung, ignoring ^C and ^Z.
This change allows user interruption of file reads during editor startup,
while editing the "next" file and when reading a file into the buffer.
If the file read is interrupted half way, no lines are put into the buffer.

Also, flush the output after printing the name of the file being read
so that they know what it is doing.