Simplify marks, allow only lower case marks and let them set ' and `
For a mark, we only need to store the line, not a position.
Instead of havving an array of 10 marks, this allows all 26 lower case
marks.
The user can also now set the previous context mark (named ' and `)
explicitly, same as classic vi.
Faster, you get all 26 marks and saves 120 bytes of code.