Don't map special keys into vi commands; add K_DELETE
Arrow keys, page up/down and so on were mapped into vi commands
like H, K, ^B, ^F in all modes, meaning that if you pressed them
in insert/replace/cmdline mode, either the command letters or
octal values from \200 upwards were inserted.
This keeps them as key codes and reacts to them or ignores them
according to the current mode.

Also, this adds recognition of the delete-character key and makes it
delete the character under the cursor when possible.