Use bc and nd event if they are more than one character long
Xvi used to use these to do relative motions only if they were a
single character. This makes it use them anyway, even if several chars.

The calculations to optimise the number of output characters for a
cursor motion still assume that moving left and down one char are
one char long, but that seems better than not using them.

This also introduced char *up, similar to bc, nd and down, for
future use.