Put cursor back on screen after "Unknown mark" message
Instead of putting the cursor back on the screen after every
error message, this makes sure the cursor position is correct
after every INSERT/REPLACE/NORMAL command.

This was triggered by "Unknown mark" leaving the cursor on the
status line, but this is a kind of error that has happened lots
of times. This comit removes all those other wind_goto()s and
puts one-for-all in the event loop.