New cursor-motion-optimisation code
This replaces the guts of xyupdate(), enables it to move right
by repainting the characters already on the screen as well as
using the UP LE RI DO capabilities to move several squares at once.

The character output for hanoi drops from 63302 to 58802 but the
code size increases from 99610 to 100798.