Get "dw" right at end of line
The fix db4b78ff to issue 21 had, as a side effect, that deleting the last
word of a line that is followed by a non-word character would also delete
that non-word character. This improves the test for deleting the last word
of the file in a way that doesn't affect last words of non-last-lines.