Revert c589e319 map-stuffed-text because it breaks hanoi
A fix for the maze-solving test, to put text inserted as input from
yank buffers through the map, broke the towers-of-hanoi-solving test
which is odd because nvi and vim, which seem to put stuffed input
through the maps, both manage to run both tests. However, for 2.50
I'd rather have a working hanoi than non-working both.