Remove limit on command line length
The command line used to be linited to WIDTH-2 characters but now
it flops from 78-cher page to 78-char page with a < continuation mark
at the LHS is you're not looking at the start of the line.

Thanks to John Vogel github.com/jvvv for the work this patch
is based on.