Fix map-stuffed-text
Chars stuffed as part of a macro expansion were being added at the end
of the non-empty canonical queue, whereas they should be stuffed at the
beginning to maintain the input in the correct order.

This also adds some tests to newline-in-stuff to ensure that multiple
lines between the first and last text fragments are inserted in the
correct order.