Motion "tc" should not move if the following char is c
xvi was written to move right one char before doing a "t" motion so that
"we can repeat searches in the same direction and have it work right."

This turns out not to be correct: Doing a "t" motion when the very next
character is the searched-for character should result in no motion,
while dtc, ytc and ctc in the same conditions should yank the one char
under the cursor.

This was the last discrepancy that stopped the maze test from working.