Fix cursor position on displayed wide characters
When the cursor is on a character like ^X or \206, it should be
displayed on the last character of its representation in normal
mode, and on the first character if in insert or replace mode.

This change fixes its position in normal mode, where it used to
be displayed on the first characters (the ^).