Fix bug introduced by ff80b4fec braking redo of an insertion
"ix[esc]." stopped working because it's stuffed in vi mode but
the last inserted text in "< doesn't want to go through the maps.

I thought that all vi-mode stuffs would be mapped and all ex-mode not
but the redo of an insertion with the . command needs to stuff in vi
mode but doesn't go through the mapping tables.

This change separates the vi_mode flag into two: vi_mode and map_it
to handle all the different cases.