Unite i_proc and r_proc as a function ir_proc with a boolean parameter
i_proc and r_proc have a lot of common code and some fixes had been
applies to one but not the other.
This turns th them into a one function, ir_proc(), which takes a
single parameter to say whether we're inserting or replacing.

This also removes the left and right arrow keys in Replace mode,
which mask some accented characters on DOS and didn't work properly.

The code saving is 656 bytes (with clang -Os for i386)