Tweak :@'s newline-adding behaviour and add "posix" parameter
POSIX says that all lines of a :@ paste should have a newline added
to them except for the last one, wich is silly because it leaves you
with the last line of the buffer hanging on the command line
waiting for you to press  [Enter].

Classic vi and vim new-line-terminates every line of a :@ or :*.
and so do we by default.

This all introduces a new parameter "posix", auto-set if
$POSIXLY_CORRECT is set, which makes it not terminate the last line.
Why you'd want that is beyond me.