Fix segfault and memory waste in cmd_types[]
Reported-by: C.J.Wagenius <cjw@tuxxor.net>

There's a problem with the cmd_types-array in cmdtab.c. It's size is set
to 256 (NUM_VI_CMD), but not all indexes are used. Many will be zeroed
out. Boundary checks are also made to NUM_VI_CMD, so handling values >
156 and < 256 will end up in a NULLified index and cause a segmentation
fault. I accidently pressed 'ä' (first byte 195) on my swedish keyboard
and it crashed.